requirements.txt
pandas>=2.2.0
openpyxl>=3.1.2
rapidfuzz>=3.9.0
Unidecode>=1.3.8
pypdf>=5.0.0
python-docx>=1.1.2
